% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{backg}
\alias{backg}
\title{A data set containing environmental conditions of background points}
\format{
A tibble object with 5000 rows and 10 variables:
\describe{
  \item{pr_ab}{background point denoted by 0}
  \item{x y}{columns with geographical coordinates}
  \item{from column aet to landform}{columns with values of environmental variables at coordinate locations}
  ...
}
}
\usage{
backg
}
\description{
A data set containing environmental conditions of background points
}
\examples{
\donttest{
require(dplyr)
data("backg")
backg
}
}
\keyword{datasets}
