% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_partition_pa.R
\name{block_partition_pa}
\alias{block_partition_pa}
\title{Spatial block cross validation}
\usage{
block_partition_pa(
  env_layer,
  data,
  x,
  y,
  pr_ab,
  n_part = 2,
  min_res_mult = 2,
  max_res_mult = 200,
  num_grids = 30
)
}
\arguments{
\item{env_layer}{raster. Raster stack or brick with environmental variable. This will be used to evaluate spatial autocorrelation and environmental similarity between training and testing partition}

\item{data}{data.frame frame with presences records}

\item{n_part}{}

\item{max_res_mult}{numeric. Maximum value will multiply raster resolution and will define the coarsest resolution to be tested, default 50.}

\item{num_grids}{numeric. Number of grid to be tested between 2x(raster resolution) and max_res_mult*(raster resolution), default 30}
}
\value{

}
\description{
Spatial block cross validation
}
