% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_psp.R
\name{data_psp}
\alias{data_psp}
\title{Calculate data to construct partial dependence surface plots}
\usage{
data_psp(
  model,
  predictors,
  resolution = 50,
  training_data = NULL,
  pchull = FALSE,
  projection_data = NULL,
  clamping = FALSE
)
}
\arguments{
\item{model}{A model object of class "gam", "gbm", "glm", "graf", "ksvm", "ksvm", "maxnet”,
“nnet", and "randomForest" This model can be found in the first element of the list returned
by any function from the fit_, tune_, or esm_ function families}

\item{predictors}{character. Vector with two predictor name(s) to plot. If NULL all predictors will
be plotted. Default NULL}

\item{resolution}{numeric. Number of equally spaced points at which to predict continuous predictors. Default 50}

\item{training_data}{data.frame. Database with response (0,1) and predictor values used
to fit a model. Default NULL}

\item{pchull}{logical. Extract convex-hull limit of training data. Default FALSE}

\item{projection_data}{SpatRaster. Raster layer with environmental variables used for model
projection. Default NULL}

\item{clamping}{logical. Perform clamping. Only for maxent models. Default FALSE}
}
\value{
A list with two tibbles "pdpdata" and "resid".
\itemize{
\item pspdata: has data to construct partial dependence surface plot, the first two column includes
values of the selected environmental variables, the third column with predicted suitability.
\item pchull: has data to plot residuals a convex hull polygon bounding calibration data.
}
}
\description{
Calculate data to construct Partial dependence surface plot (i.e., bivariate dependence plot) for two predictor set
}
\examples{
\dontrun{
library(terra)
library(dplyr)

somevar <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(somevar) # environmental data
names(somevar) <- c("aet", "cwd", "tmx", "tmn")
data(abies)

abies2 <- abies \%>\%
  select(x, y, pr_ab)

abies2 <- sdm_extract(abies2,
                      x = "x",
                      y = "y",
                      env_layer = somevar
)
abies2 <- part_random(abies2,
                      pr_ab = "pr_ab",
                      method = c(method = "kfold", folds = 5)
)

m <- fit_svm(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "cwd", "tmx", "tmn"),
  partition = ".part",
  thr = c("max_sens_spec")
)

df <- data_psp(
  model = m$model,
  predictors = c("aet", "cwd"),
  resolution = 50,
  projection_data = somevar,
  pchull = TRUE,
  training_data = abies2,
  clamping = TRUE
)

df
names(df)
df$pspdata
df$pchull

# see p_psp to construct partial dependence plot with ggplot2
}
}
\seealso{
\code{\link{data_pdp},  \link{p_psp}, \link{p_pdp}}
}
