% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_filtering.R
\name{env_filtering}
\alias{env_filtering}
\title{Function to perform environmental filtering of species occurrence records}
\usage{
env_filtering(da, x, y, id, variables, nbins, cores = 1, plot = TRUE)
}
\arguments{
\item{variables}{data.frame. A data.frame with environmental conditions. It is possible use two or three variables.}

\item{nbins}{numeric. A number of classes used to split each environmental condition.}

\item{plot}{logical. Plot filtering procedure.}

\item{coord}{data.frame. A data.frame with longitude and latitude in the first and second columns.}
}
\description{
Function to perform environmental filtering of species occurrence records
}
