% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_exclude.R
\name{extra_exclude}
\alias{extra_exclude}
\title{Constraint of suitability based on extrapolation}
\usage{
extra_exclude(suit, extra, threshold = 50)
}
\arguments{
\item{suit}{SpatRaster with suitability values}

\item{extra}{SpatRaster with extrapolation values measured in percentage (output from extra_eval function)}

\item{threshold}{numeric. Vector with one or more values used for correct extrapolation. Default 50\% (DOES THIS FUNCTION SET THE PROJECTED SUITABILITY VALUES LESS THAN THE THRESHOLD TO ZERO? UNCLEAR. PLEASE BE EXPLICIT)}
}
\value{
A SpatRaster object with corrected suitability values
}
\description{
Exclusion of suitability values less than a given extrapolation value (EXPERIMENTAL)
}
\examples{
\dontrun{
# see examples in extra_eval function
}
}
\seealso{
\code{\link{extra_eval}}
}
