% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_truncate.R
\name{extra_truncate}
\alias{extra_truncate}
\title{Truncate suitability predictions based on an extrapolation value}
\usage{
extra_truncate(suit, extra, threshold = 50, trunc_value = 0)
}
\arguments{
\item{suit}{SpatRaster with suitability values}

\item{extra}{SpatRaster with extrapolation values preferable measured with extra_eval function}

\item{threshold}{numeric. Vector with one or more extrapolation values used for truncate suitability Default 50\%}

\item{trunc_value}{numeric. Numeric value to be used to those cells assumed to be extrapolative}
}
\value{
A SpatRaster object with truncated suitability values
}
\description{
Exclusion of suitability predictions in environmental conditions assumed to be
extrapolative.
}
\details{
Exclusion of suitability predictions in environmental conditions assumed to be
extrapolative. In this function it is possible to use any metric measuring degree of
extrapolation (e.g., MESS-Multivariate Environmental Similarity Surfaces, EO-Environmental
Overlap, MOP-Mobility-Oriented Parity, EXDET-Extrapolation Detection, or AOA-Area of
Applicability). However, we recommend to use Shape approach (see \code{\link{extra_eval}},
and Velazco et al., 2023; \doi{10.1111/ecog.06992}).

This function truncates suitability predictions assigning a given value, generally 0 or NA.
 Usage trunc_value = NA. Default 0.

To those cells assumed to be extrapolative, i.e., higher than a given threshold of a given
extrapolation metric.

See this \href{https://sjevelazco.github.io/flexsdm/articles/v06_Extrapolation_example.html}{vignette at flexsdm website}
for further details about Shape metric, model truncation, and tools to explore model extrapolation.
}
\examples{
\donttest{
# see examples in extra_eval function
}
}
\seealso{
\code{\link{extra_eval}}, \code{\link{p_extra}}, \code{\link{p_pdp}}, \code{\link{p_bpdp}}
}
