% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mx.R
\name{fit_mx}
\alias{fit_mx}
\title{Fit and validate Maximum Entropy models}
\usage{
fit_mx(
  data,
  response,
  predictors,
  predictors_f = NULL,
  partition,
  background,
  thr = NULL,
  fit_formula = NULL,
  clamp = TRUE,
  pred_type = "cloglog",
  ...
)
}
\arguments{
\item{data}{data.frame. Database with response (0,1) and predictors values.}

\item{response}{character. Column name with species absence-presence data (0,1).}

\item{predictors}{character. Vector with the column names of quantitative
predictor variables (i.e. continuous or discrete variables).
Usage predictors = c("aet", "cwd", "tmin")}

\item{predictors_f}{character. Vector with the column names of qualitative
predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{partition}{character. Column name with training and validation partition groups.}

\item{background}{data.frame. Database with response column only with 0 and predictors variables. All
column names must be consistent with data}

\item{thr}{character. Threshold used to get binary suitability values (i.e. 0,1).
It is useful for threshold-dependent performance metrics.
It is possible to use more than one threshold type. It is necessary to provide a
vector for this argument. The next threshold area available:
\itemize{
  \item LPT: The highest threshold at which there is no omission. Usage thr=c(type='LPT').
  \item EQUAL_SENS_SPEC: Threshold at which the sensitivity and specificity are equal.
  \item MAX_TSS: Threshold at which the sum of the sensitivity and specificity is the highest.
  Usage thr=c(type='MAX_TSS').
  \item MAX_KAPPA: The threshold at which kappa is the highest ("max kappa"). Usage thr=c(type='MAX_KAPPA').
  \item MAX_JACCARD: The threshold at which Jaccard is the highest. Usage thr=c(type='MAX_JACCARD').
  \item MAX_SORENSEN: The threshold at which Sorensen is highest. Usage thr=c(type='MAX_SORENSEN').
  \item MAX_FPB: The threshold at which FPB is highest. Usage thr=c(type='MAX_FPB').
  \item SENSITIVITY: A threshold value specified by user. Usage thr=c(type='SENSITIVITY', sens='0.6'). 'sens' refers to models will be binarized using this suitability value.
  }}

\item{fit_formula}{formula. A formula object with response and predictor
variables see maxnet.formula function from maxnet package.
Note that the variables used here must be consistent with those used in
response, predictors, and predictors_f arguments}

\item{clamp}{logical. It is set with TRUE, predictors and features are restricted to the range seen during model training.}

\item{pred_type}{character. Type of response required availabe "link", "exponential", "cloglog" and "logistic". Dafault "cloglog"}

\item{...}{}
}
\value{
A list object with:
\itemize{
\item model: A "ksvm" class object. This object can be used for predicting.
\item performance: Performance metric (see \code{\link{enm_eval}}).
Those threshold dependent metric are calculated based on the threshold specified in thr argument .
\item selected_threshold: Value of the threshold selected.
\item threshold_table: Value of all threshold.
}
}
\description{
Fit and validate Maximum Entropy models
}
\examples{
\dontrun{
}

}
