% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hespero}
\alias{hespero}
\title{A data set containing localities of Hesperocyparis stephensonii species in California, USA}
\format{
A tibble object with 14 rows and 4 variables:
\describe{
  \item{ID}{presences records ID}
  \item{x y}{columns with coordinates in Albers Equal Area Conic coordinate system}
  \item{pr_ab}{presence denoted by 1}
  ...
}
}
\usage{
hespero
}
\description{
A data set containing localities of Hesperocyparis stephensonii species in California, USA
}
\examples{
\donttest{
require(dplyr)
data("hespero")
hespero
}
}
\keyword{datasets}
