% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogenize_na.R
\name{homogenize_na}
\alias{homogenize_na}
\title{Homogenize cells with NAs across all layers}
\usage{
homogenize_na(x)
}
\arguments{
\item{x}{A SpatRaster.}
}
\value{
a SpatRaster
}
\description{
Homogenize cells with NAs across all layers
}
\details{
Homogenize cells with NAs across all layers in a SpatRaster so that in the resulting SpatRaster all layers have the same cells with NAa
}
\examples{
\donttest{
require(terra)

somevar <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(somevar)

somevar2 <- homogenize_na(somevar)
par(mfrow = c(2, 1))
plot(somevar$CFP_4)
plot(somevar2$CFP_4)
par(mfrow = c(1, 1))

# In somevar2 all layers have the same cells with NAs
}
}
