% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palms}
\alias{palms}
\title{A data set containing presences of palms species from Southern Brazil}
\format{
A tibble with 327 rows and 3 variables:
\describe{
  \item{species}{species names}
  \item{x}{longitude of species occurrences}
  \item{y}{latitude of species occurrences}
  ...
}
}
\usage{
palms
}
\description{
This data set contains presences of 11 palm species from Southern Brazil sourced by Calambás-Trochez et al. (2021).
}
\examples{
\donttest{
require(dplyr)
data("palms")
}
}
\references{
Calambás-Trochez, L.F., Velazco, S.J.E., Hoffmann, P.M., Brum, F.T., Carlucci, M.B., 2021. Climate and land-use changes coupled with low coverage of protected areas threaten palm species in South Brazilian grasslands. Perspectives in Ecology and Conservation 9. https://doi.org/10.1016/j.pecon.2021.03.010
}
\keyword{datasets}
