% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part.R
\name{part}
\alias{part}
\title{Data partitioning for training and testing models}
\usage{
part(data, pr_ab, bg_data = NULL, bg_a = NULL, method = NULL)
}
\arguments{
\item{data}{data.frame. Database with presences, presence-absence, or pseudo-absence, records for a given species}

\item{pr_ab}{character. Column name of "data" with presences, presence-absence, or pseudo-absence. Presences must be represented by 1 and absences by 0}

\item{bg_data}{data.frame. Data frames with background points.}

\item{bg_a}{character. Column name of "bg_data" with absences. Background must be repented by 0}

\item{method}{character. Vector with partition method to be used:
\itemize{
  \item kfold: Random partition in k-fold cross-validation. Usage part=c(method= 'kfold', folds='5'). 'folds' refers to the number of folds for data partitioning, it assumes value >=1. Usage method = c(method = "kfold", folds = 10).
  \item rep_kfold: Random partition in repeated k-fold cross-validation. Usage method = c(method = "rep_kfold", folds = 10, replicates=10). 'folds' refers to the number of folds for data partitioning, it assumes value >=1. 'replicate' refers to the number of replicates, it assumes a value >=1.
  \item loocv: Leave-one-out cross-validation (a.k.a. Jackknife) WRITE METHOD HERE . Usage method = c(method = "loocv").
  \item boot: Random bootstrap partition. Usage part=c(method='boot', replicates='2', proportion='0.7'). 'replicate' refers to the number of replicates, it assumes a value >=1. 'proportion' refers to the proportion of occurrences used for model fitting, it assumes a value >0 and <=1. In this example proportion='0.7' mean that 70% of data will be used for model training, while 30% for model testing. For this method function will returns .partX columns with "train" and "test" words
  }}
}
\value{
A tibble object with information used in 'data' and/or 'gb_data' arguments and a additional columns .part with partition group. In the case of boot partition partition data will be stored with 'train' and 'test' words, for all other methods partition groups will be stored as numeric. In case of use 'data' and 'bg_data' will returned a list with partition data for each database with the names data and bg_data.
}
\description{
This function provides different conventional partition methods based in folds (kfold, rep_kfold, and loocv), and bootstrap (boot)
}
\examples{
\dontrun{
data("abies_db")
abies_db$partition <- NULL
abies_db <- tibble(abies_db)

# K-fold method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  bg_data = NULL,
  bg_a = NULL,
  method = c(method = "kfold", folds = 10)
)
abies_db2

# Repeated K-fold method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  bg_data = NULL,
  bg_a = NULL,
  method = c(method = "rep_kfold", folds = 10, replicates = 10)
)
abies_db2

# Leave-one-out cross-validation (loocv) method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  bg_data = NULL,
  bg_a = NULL,
  method = c(method = "loocv")
)
abies_db2

# Bootstrap method
abies_db2 <- part(
  data = abies_db,
  pr_ab = "pr_ab",
  bg_data = NULL,
  bg_a = NULL,
  method = c(method = "boot", replicates = 50, proportion = 0.7)
)
abies_db2
abies_db2$.part1 \%>\% table() # Note that for this method .partX columns have train and test words.
}

}
\seealso{
\code{\link{part_sblock}}, \code{\link{part_senv}}, \code{\link{sample_pseudoabs}}, \code{\link{sample_background}}
}
