% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_predict_test.R
\name{sdm_predict_test}
\alias{sdm_predict_test}
\title{Spatial model prediction and ensemble EXPERIMENTAL}
\usage{
sdm_predict_test(
  models,
  pred,
  thr,
  calib_area = NULL,
  ensemble = NULL,
  clamp = TRUE,
  pred_type = "cloglog"
)
}
\arguments{
\item{models}{list. A list a single or several models fitted with some of fit_ or tune_ functions}

\item{pred}{stack or brick. Raster layer with predictor variables. Names of raster layers must exactly match those used in model fitting.}

\item{thr}{character. Binarize projection. Default NULL, i.e. function returns only continuous projection. If used with 'selected_thr', function returns continuous and binarized models used in the 'thr' argument of some of fit_ or tune_ functions. It used with "all_thr" , function returns continuous and binarized for all threshold types.}

\item{calib_area}{SpatialPolygon or SpatialPolygonDataFrame. Spatial polygon used for restrinc prediction into a given region. Default = NULL}

\item{ensemble}{character. Method used to ensemble different models It is possible to use more than one method. A vector must be provided for this argument. For meansup, meanw or pcasup method it is necessary provide an evaluation metric to ensemble arguments (i.e., AUC, Kappa, TSS, Jaccard, Sorensen or Fpb) see below. (default NULL):
  \itemize{
  \item mean: Simple average of the different models. Usage ensemble=c(method='mean').
  \item meanw: Weighted average of models based on their performance. An evaluation metric and threshold type must be provided. Usage ensemble = c(method='meanw', metric='TSS', thr = 'max_sens_spec').
  \item meansup: Average of the best models (e.g., TSS over the average). An evaluation metric must be provided. Usage ensemble=c(method='meansup', metric='TSS').
  \item meanthr: pca performed only with those cells with suitability values above the selected threshold. Usage ensemble=c(method='pcathr').
  \item median: Median of the different models. Usage ensemble = c(method = 'median')
  }

 In the case of use more than one ensemble method it is necessary concatenate the names of ensemble methods within the argument, e.g., ensemble=c(method=c('mean', 'median')) or ensemble = c(method=c('mean', 'meanw', 'meansup', 'meanthr', 'median'), metric='TSS', thr = 'max_sens_spec')}

\item{clamp}{logical. It is set with TRUE, predictors and features are restricted to the range seen during model training. Only valid for Maxent model (see tune_mx and fit_mx)}

\item{pred_type}{character. Type of response required available "link", "exponential", "cloglog" and "logistic". Default "cloglog". Only valid for Maxent model (see tune_mx and fit_mx)}
}
\value{
A list of Raster or RasterStack with continuous or continuous and binary predictions
}
\description{
This function allows the prediction of one or more models built with the fit_ or tune_ function set. It can return continuous or continuous and binary predictions for one or more thresholds
}
