% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_mx.R
\name{tune_mx}
\alias{tune_mx}
\title{Function for constructing Maximum Entropy with exploration of hyper-parameters}
\usage{
tune_mx(
  data,
  response,
  predictors,
  predictors_f = NULL,
  background = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  clamp = TRUE,
  pred_type = "cloglog",
  ...
)
}
\arguments{
\item{pred_type}{}
}
\value{

}
\description{
Function for constructing Maximum Entropy with exploration of hyper-parameters
}
