% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_nnet.R
\name{tune_nnet}
\alias{tune_nnet}
\title{Fit and validate Neural Networks models with exploration of hyper-parameters}
\usage{
tune_nnet(
  data,
  response,
  predictors,
  predictors_f = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  ...
)
}
\arguments{
\item{metric}{}
}
\value{

}
\description{
Fit and validate Neural Networks models with exploration of hyper-parameters
}
