% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_rf.R
\name{tune_rf}
\alias{tune_rf}
\title{Fit and validate Random Forest models with exploration of hyper-parameters}
\usage{
tune_rf(
  data,
  response,
  predictors,
  predictors_f = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  ...
)
}
\arguments{
\item{metric}{}
}
\value{

}
\description{
Fit and validate Random Forest models with exploration of hyper-parameters
}
