% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abies}
\alias{abies}
\title{A data set containing localities and environmental condition of an Abies (fir tree) species in California, USA}
\format{
A tibble object with 5000 rows and 10 variables:
\describe{
  \item{ID}{presences and absences records ID}
  \item{pr_ab}{presence and absences denoted by 1 and 0 respectively}
  \item{x y}{columns with coordinates in Albers Equal Area Conic coordinate system}
  \item{from column aet to landform}{columns with values for environmental variables at each locality}
  ...
}
}
\usage{
abies
}
\description{
A data set containing localities and environmental condition of an Abies (fir tree) species in California, USA
}
\examples{
\dontrun{
require(dplyr)
data("abies")
abies
}
}
\keyword{datasets}
