% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abies}
\alias{abies}
\title{A data set containing environmental condition of an Abies species}
\format{
A tibble object with 5000 rows and 10 variables:
\describe{
  \item{ID}{presence and absences records ID}
  \item{pr_ab}{presence and absences denoted by 1 and 0 respectively}
  \item{x y}{columns with coordinates}
  \item{from column aet to landform}{columns with environmental variables}
  ...
}
}
\usage{
abies
}
\description{
A data set containing environmental condition of an Abies species
}
\examples{
\dontrun{
require(dplyr)
data("abies")
abies
}
}
\keyword{datasets}
