% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{backg}
\alias{backg}
\title{A data set containing environmental conditions of background points}
\format{
A tibble object with 5000 rows and 11 variables:
\describe{
  \item{pr_ab}{presence and absences denoted by 1 and 0 respectively}
  \item{from column aet to landform}{columns with environmental variables}
  ...
}
}
\usage{
backg
}
\description{
A data set containing environmental conditions of background points
}
\examples{
\dontrun{
require(dplyr)
data('backg')
spp
}
}
\keyword{datasets}
