% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_partition_pa.R
\name{block_partition_pa}
\alias{block_partition_pa}
\title{Title}
\usage{
block_partition_pa(
  env_layer,
  occ_data,
  sp,
  x,
  y,
  pr_ab,
  min_res_mult = 2,
  max_res_mult = 200,
  num_grids = 30,
  n_part = 2,
  cores = 1,
  save_part_raster = FALSE,
  dir_save
)
}
\arguments{
\item{env_layer}{}

\item{occ_data}{}

\item{sp}{}

\item{x}{}

\item{y}{}

\item{pr_ab}{}

\item{min_res_mult}{}

\item{max_res_mult}{}

\item{num_grids}{}

\item{n_part}{}

\item{cores}{}

\item{save_part_raster}{}

\item{dir_save}{}
}
\value{

}
\description{
Title
}
