% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_colinvar.R
\name{correct_colinvar}
\alias{correct_colinvar}
\title{Perform collinearity reduction on predictors}
\usage{
correct_colinvar(rstack, method, proj = NULL)
}
\arguments{
\item{rstack}{SpatRaster An object of class SpatRaster containing the predictors.}

\item{method}{character. Collinearity reduction method. It is necessary to
provide a vector for this argument. The next methods are implemented:
\itemize{
  \item pearson: Select variables by Pearson correlation, a threshold of maximum correlation must be specified by user. Otherwise a threshold of 0.7 is defined as default. Usage method = c('pearson', th='0.7').
  \item vif: Select variables by Variance Inflation Factor, a threshold can be specified by user. Otherwise a threshold of 10 is defined as default.Usage method = c('vif', th = '10').
  \item pca: Perform a Principal Component Analysis and use the principal components as the new predictors. The selected components account for 95% of the whole variation in the system.Usage method = c('pca').
  \item fa: Perform a Factorial Analysis and select, from the original predictors, the ones with the highest correlation with the axis which account for 95% of the whole variation in the system.Usage method = c('fa').
}}

\item{proj}{character. Path to a folder which contains sub-folders for the different projection scenarios. Only used for pca. Usage "C:/User/Desktop/Projections"}
}
\value{

}
\description{
Perform collinearity reduction on predictors
}
\examples{
\dontrun{
somevar <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(somevar)

# Perform pearson collinearity control
somevar <- correct_colinvar(rstack = somevar, method = c("pearson", th = "0.8"))
somevar$rstack
somevar$removed_variables
somevar$correlation_table

# Perform vif collinearity control
somevar <- correct_colinvar(rstack = somevar, method = c("vif", th = "8"))
somevar$rstack
somevar$removed_variables
somevar$correlation_table

# Perform pca collinearity control
somevar <- correct_colinvar(rstack = somevar, method = c("pca"))
somevar$rstack
somevar$coeficients
somevar$cumulative_variance

# Perform pca collinearity control for projections
somevar <- correct_colinvar(rstack = somevar, method = c("pca", proj = ))
somevar$rstack
somevar$coeficients
somevar$cumulative_variance
somevar$proj

# Perform fa collinearity control
somevar <- correct_colinvar(rstack = somevar, method = c("fa"))
somevar$rstack
somevar$removed_variables
somevar$correlation_table
}

}
