% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_filtering.R
\name{env_filtering}
\alias{env_filtering}
\title{Function to perform environmental filtering of species occurrence records}
\usage{
env_filtering(data, x, y, id, variables, nbins, cores = 1)
}
\arguments{
\item{data}{data.frame. Data.frame or tibble object with presences
(or presence-absence) records, and coordinates}

\item{x}{character. Column name with longitude data}

\item{y}{character. Column name with latitude data}

\item{id}{character. Column names with rows id. It is important that each row has its own unique code.}

\item{variables}{data.frame. A data.frame with environmental conditions. It is possible use two or three variables}

\item{nbins}{interger. A number of classes used to split each environmental condition}

\item{cores}{interger. Number of machine cores used for processing in parallel}
}
\description{
Function to perform environmental filtering of species occurrence records
}
