% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_outliers.R
\name{env_outliers}
\alias{env_outliers}
\title{Detection of outliers in the environmental space}
\usage{
env_outliers(da, species, x, y, pr_ab, id, envr)
}
\description{
Detection of outliers in the environmental space
}
