% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_eval.R
\name{extra_eval}
\alias{extra_eval}
\title{Measure model extrapolation}
\usage{
extra_eval(env_calib, env_proj)
}
\arguments{
\item{env_calib}{SpatRaster with environmental conditions of the calibration area or the
presence and absence points localities used for constructing models}

\item{env_proj}{SpatRaster with environmental condition used for projecting a model}
}
\value{
A SpatRaster object with extrapolation values measured in percentage
}
\description{
Measure model extrapolation
}
\seealso{
\code{\link{extra_correct}}
}
