% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_truncate.R
\name{extra_truncate}
\alias{extra_truncate}
\title{Truncate suitability predictions based on an extrapolation value}
\usage{
extra_truncate(suit, extra, threshold = 50, trunc_value = 0)
}
\arguments{
\item{suit}{SpatRaster with suitability values}

\item{extra}{SpatRaster with extrapolation values preferable measured with extra_eval function}

\item{threshold}{numeric. Vector with one or more extrapolation values used for truncate suitability Default 50\%}

\item{trunc_value}{numeric. Numeric value to be used to those cells assumed to be extrapolative}
}
\value{
A SpatRaster object with truncated suitability values
}
\description{
Exclusion of suitability predictions in environmental conditions assumed as with high extrapolation value (EXPERIMENTAL)
}
\details{
This function truncates suitability predictions assigning a given value, generally 0 or NA. Usage trunc_value = NA. Default 0.
to those cells assumed to be extrapolative (i.e., higher than a given threshold) based on an extrapolation metric like SHAPE methods (calculated with extra_eval).
We recommend using the function p_pdp and p_extra.
}
\examples{
\dontrun{
# see examples in extra_eval function
}
}
\seealso{
\code{\link{extra_eval}}
}
