% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ensemble.R
\name{fit_ensemble}
\alias{fit_ensemble}
\title{Model assemble and validation}
\usage{
fit_ensemble(
  models,
  ens_method = c("mean", "meanw", "meansup", "meanthr", "median"),
  thr,
  metric = "TSS"
)
}
\arguments{
\item{models}{list. A list of models fitted with fit_ or tune_ function family. Models used for ensemble must have the same presences-absences records, partition methods, threshold types.}

\item{thr}{character. Threshold type used to get binary suitability values (i.e. 0,1). In fit_ensemble, only it possible to use one threshold type. It is mandatory that the threshold used was the same for all models (see thr argument in fit_ and tune_ family function). Threshold available: lpt, equal_sens_spec, max_sens_spec, max_kappa, max_jaccard, max_sorensen, max_fpb, and sensitivity. Usage thr = 'equal_sens_spec'}

\item{metric}{character. Performance metric used for selecting the best combination of hyper-parameter values. One of the next metrics can be used SORENSEN, JACCARD, FPB, TSS, KAPPA, AUC, IMAE, and BOYCE. Default TSS. Usage metric = BOYCE}

\item{ensemble}{character. Method used to ensemble different models. A vector must be provided for this argument. For meansup, meanw or pcasup method, it is necessary to provide an evaluation metric to ensemble arguments (metric argument below). By default will be performed all ensemble methods:
\itemize{
\item mean: Simple average of the different models. Usage ensemble=c(method='mean').
\item meanw: Weighted average of models based on their performance. An evaluation metric and threshold type must be provided..
\item meansup: Average of the best models (e.g., TSS over the average). An evaluation metric must be provided.
\item meanthr: Average performed only with those cells with suitability values above the selected threshold.
\item median: Median of the different models.
}
Usage ensemble = "meanthr". In case is needed perform different ensemble method it is necessary concatenate them, e.g., ensemble = c("meanw", "meanthr", "median")}
}
\value{
A list object with:
\itemize{
\item model: A list of models used for performing ensemble.
\item predictors: A tibble of quantitative (column names with c) and qualitative (column names with f) variables used in each models.
\item performance: A tibble with performance metric (see \code{\link{sdm_eval}}).
Those threshold dependent metrics are calculated based on the threshold specified in thr argument .
\item selected_thresholds: A tibble with value of the threshold selected.
\item threshold_table: A tibble with all threshold values.
}
}
\description{
Model assemble and validation
}
