% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gbm.R
\name{fit_gbm}
\alias{fit_gbm}
\title{Fit and validate Generalized Boosted Regression models}
\usage{
fit_gbm(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  thr = NULL,
  n_trees = 100,
  n_minobsinnode = 10,
  shrinkage = 0.1,
  ...
)
}
\arguments{
\item{data}{data.frame. Database with response (0,1) and predictors values.}

\item{response}{character. Column name with species absence-presence data (0,1).}

\item{predictors}{character. Vector with the column names of quantitative
predictor variables (i.e. continuous or discrete variables).
Usage predictors = c("aet", "cwd", "tmin")}

\item{predictors_f}{character. Vector with the column names of qualitative
predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor
variables (e.g. formula(pr_ab ~ aet + ppt_jja + pH + awc + depth + landform)).
Note that the variables used here must be consistent with those used in
response, predictors, and predictors_f arguments. Default is NULL.}

\item{partition}{character. Column name with training and validation partition groups.}

\item{thr}{character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:
\itemize{
  \item lpt: The highest threshold at which there is no omission. Usage thr=c(type='lpt').
  \item equal_sens_spec: Threshold at which the sensitivity and specificity are equal (aka threshold that maximizes the TSS).
  \item max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest.
  Usage thr=c(type='max_sens_spec').
  \item max_kappa: The threshold at which Kappa is the highest ("max kappa"). Usage thr=c(type='max_kappa').
  \item max_jaccard: The threshold at which Jaccard is the highest. Usage thr=c(type='max_jaccard').
  \item max_sorensen: The threshold at which Sorensen is highest. Usage thr=c(type='max_sorensen').
  \item max_fpb: The threshold at which FPB is highest. Usage thr=c(type='max_fpb').
  \item specific: A threshold value specified by user. Usage thr=c(type='specific', sens='0.6'). 'sens' refers to models will be binarized using this suitability value.
  }}

\item{n_trees}{Integer specifying the total number of trees to fit.
This is equivalent to the number of iterations and the number of basis
functions in the additive expansion. Default is 100.}

\item{n_minobsinnode}{Integer. It specifies the minimum number of
observations in the terminal nodes of the trees. Note that this
is the actual number of observations, not the total weight.}

\item{shrinkage}{Numeric. This parameter applied to each tree in the
expansion. Also known as the learning rate or step-size reduction;
0.001 to 0.1 usually work, but a smaller learning rate typically
requires more trees. Default is 0.1.}

\item{...}{}
}
\value{
A list object with:
\itemize{
\item model: A "gbm" class object. This object can be used for predicting.
\item predictors: A character with quantitative (elements names with c) and qualitative (elements names with f) variables use for modeling.
\item performance: Performance metric (see \code{\link{sdm_eval}}).
Those threshold dependent metric are calculated based on the threshold specified in thr argument .
\item selected_thresholds: Value of the threshold selected.
\item threshold_table: Value of all threshold.
}
}
\description{
Fit and validate Generalized Boosted Regression models
}
\examples{
\dontrun{
data("abies_db")

# Using k-fold partition method
abies_db2 <- data_part(
  data = abies_db,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 10)
)
abies_db2

gbm_t1 <- fit_gbm(
  data = abies_db2,
  response = "pr_ab",
  predictors = c("aet", "ppt_jja", "pH", "awc", "depth"),
  predictors_f = c("landform"),
  partition = ".part",
  thr = c("max_sens_spec", "equal_sens_spec", "mas_sorensen")
)
gbm_t1$model
gbm_t1$performance
gbm_t1$selected_thresholds
gbm_t1$all_thresholds

# Using bootstrap partition method
abies_db2 <- data_part(
  data = abies_db,
  pr_ab = "pr_ab",
  method = c(method = "boot", replicates = 10, proportion = 0.7)
)
abies_db2

gbm_t2 <- fit_gbm(
  data = abies_db2,
  response = "pr_ab",
  predictors = c("ppt_jja", "pH", "awc"),
  predictors_f = c("landform"),
  partition = ".part",
  thr = "max_sens_spec",
  poly = 3,
  inter_order = 2
)
gbm_t2
}

}
