% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_max.R
\name{fit_max}
\alias{fit_max}
\title{Fit and validate Maximum Entropy models}
\usage{
fit_max(
  data,
  response,
  predictors,
  predictors_f = NULL,
  fit_formula = NULL,
  partition,
  background = NULL,
  thr = NULL,
  clamp = TRUE,
  classes = "default",
  pred_type = "cloglog",
  regmult = 1
)
}
\arguments{
\item{data}{data.frame. Database with response (0,1) and predictors values.}

\item{response}{character. Column name with species absence-presence data (0,1).}

\item{predictors}{character. Vector with the column names of quantitative
predictor variables (i.e. continuous variables).
Usage predictors = c("aet", "cwd", "tmin")}

\item{predictors_f}{character. Vector with the column names of qualitative
predictor variables (i.e. ordinal or nominal variables type). Usage predictors_f = c("landform")}

\item{fit_formula}{formula. A formula object with response and predictor
variables see maxnet.formula function from maxnet package.
Note that the variables used here must be consistent with those used in
response, predictors, and predictors_f arguments. Default NULL.}

\item{partition}{character. Column name with training and validation partition groups.}

\item{background}{data.frame. Database with response column only with 0 and predictors variables. All
column names must be consistent with data. Default NULL}

\item{thr}{character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:
\itemize{
  \item lpt: The highest threshold at which there is no omission.
  \item equal_sens_spec: Threshold at which the sensitivity and specificity are equal.
  \item max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest (aka threshold that maximizes the TSS).
  \item max_jaccard: The threshold at which Jaccard is the highest.
  \item max_sorensen: The threshold at which Sorensen is highest.
  \item max_fpb: The threshold at which FPB is highest.
  \item sensitivity: Threshold based on a specified sensitivity value.
  Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers to sensitivity value. If it is not specified a sensitivity values, function will use by default 0.9
  }
In the case of use more than one threshold type it is necessary concatenate threshold types, e.g., thr=c('lpt', 'max_sens_spec', 'max_jaccard'), or thr=c('lpt', 'max_sens_spec', 'sensitivity', sens='0.8'), or thr=c('lpt', 'max_sens_spec', 'sensitivity'). Function will use all thresholds if no threshold is specified}

\item{clamp}{logical. It is set with TRUE, predictors and features are restricted to the range seen during model training.}

\item{classes}{character. A single feature of any combinations of them. Features are symbolized by letters: l (linear), q (quadratic), h (hinge), p (product), and t (threshold). Usage classes = "lpq". Default "default" (see details).}

\item{pred_type}{character. Type of response required available "link", "exponential", "cloglog" and "logistic". Default "cloglog"}

\item{regmult}{numeric. A constant to adjust regularization. Default 1.}
}
\value{
A list object with:
\itemize{
\item model: A "MaxEnt" class object. This object can be used for predicting.
\item predictors: A tibble with quantitative (c colum names) and qualitative (f colum names) variables use for modeling.
\item performance: Performance metric (see \code{\link{sdm_eval}}).
Those threshold dependent metric are calculated based on the threshold specified in thr argument .
\item selected_thresholds: Value of the threshold selected.
\item all_thresholds: Value of all threshold.
}
}
\description{
Fit and validate Maximum Entropy models
}
\details{
EXPLAIN HERE DEFAUL SELECTION OF FEATURES BASED ON number of occurrences
if (np < 10) {
  classes <- "l"
} else if (np < 15){
  classes <- "lq"
} else if (np < 80) {
  classes <- "lqh"
}
}
\examples{
\dontrun{
data("abies")
data("backg")
abies # environmental conditions of presence-absence data
backg # environmental conditions of background points

# Using k-fold partition method
# Remember that the partition method, number of folds or replications must
# be the same for presence-absence and background points datasets
abies2 <- part_random(
  data = abies,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 5)
)
abies2

backg2 <- part_random(
  data = backg,
  pr_ab = "pr_ab",
  method = c(method = "kfold", folds = 5)
)
backg2

max_t1 <- fit_max(
  data = abies2,
  response = "pr_ab",
  predictors = c("aet", "ppt_jja", "pH", "awc", "depth"),
  predictors_f = c("landform"),
  partition = ".part",
  background = backg2,
  thr = c("max_sens_spec", "equal_sens_spec", "max_sorensen"),
  clamp = TRUE,
  classes = "default",
  pred_type = "cloglog",
  regmult = 1
)
length(max_t1)
max_t1$model
max_t1$predictors
max_t1$performance
max_t1$data_ens
}

}
