% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block.R
\name{get_block}
\alias{get_block}
\title{Transform a spatial partition layer to the same spatial properties of environmental variables}
\usage{
get_block(env_layer, best_grid)
}
\arguments{
\item{env_layer}{raster. A raster, stack, or brick object with some environmental variables used in the block_partition or band_partition function. Function always will select the first layer}

\item{best_grid}{raster. Raster object returned by block_partition or band_partition}
}
\value{

}
\description{
Transform a spatial partition layer to the same spatial properties of environmental variables
}
\details{
Transform a layer originated from the function block_partition or band_partition to the same properties of environmental variables
}
\examples{
\dontrun{
See examples in block_partition \code{\link{block_partition}} ro \code{\link{band_partition}}
}

}
