% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter.R
\name{inter}
\alias{inter}
\title{Raster interpolation}
\usage{
inter(r1, r2, y1, y2, rastername = NULL, dir_save = NULL, n_cores = 1)
}
\arguments{
\item{r1}{raster. Raster object for the initial year}

\item{r2}{raster. Raster object for the final year}

\item{y1}{numeric. Initial year}

\item{y2}{numeric. Final year}

\item{rastername}{character. word used as prefix in raster file name}

\item{dir_save}{character. Directory path and name of the folder in which you want to save the raster files}

\item{n_cores}{numeric. Number of cores use for parallelization}
}
\value{

}
\description{
Raster interpolation
}
