% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdm_posteriori.R
\name{msdm_posteriori}
\alias{msdm_posteriori}
\title{Methods to correct overprediction of species distribution models based on occurrences and suitability patterns.}
\usage{
msdm_posteriori(
  records,
  x,
  y,
  pr_ab,
  cont_suit,
  method = c("obr", "pres", "lq", "mcp", "bmcp"),
  thr = "equal_sens_spec",
  buffer = NULL
)
}
\arguments{
\item{records}{tibble or data.frame. A database with geographical coordinates of species presences and absences (or pseudo-absence) used to create species distribution models.}

\item{x}{character. Column name with longitude values.}

\item{y}{character. Column name with latitude values.}

\item{pr_ab}{character. Column name with presence and absence data (i.e. 1 and 0)}

\item{cont_suit}{SpatRaster. Raster with continuous suitability predictions
"species_specific" type calculates the minimum pairwise-distances between all occurrences and then selects the maximum distance, i.e., the value of the buffer will be the maximum distance from the minimum distance. This procedure depends on the presence of each species' occurrences; thus, for each species, a value of buffer width will be calculated (usage buffer="species_specific").}

\item{method}{character. A character string indicating which constraint method must be used create.}

\item{thr}{character. Threshold used to get binary suitability values (i.e. 0,1). It is useful for threshold-dependent performance metrics. It is possible to use more than one threshold type. It is necessary to provide a vector for this argument. The next threshold area available:
\itemize{
  \item lpt: The highest threshold at which there is no omission.
  \item equal_sens_spec: Threshold at which the sensitivity and specificity are equal.
  \item max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest (aka threshold that maximizes the TSS).
  \item max_jaccard: The threshold at which Jaccard is the highest.
  \item max_sorensen: The threshold at which Sorensen is highest.
  \item max_fpb: The threshold at which FPB is highest.
  \item sensitivity: Threshold based on a specified sensitivity value.
  Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers to sensitivity value. If it is not specified a sensitivity values, function will use by default 0.9
  }
In the case of use more than one threshold type it is necessary concatenate threshold types, e.g., thr=c('lpt', 'max_sens_spec', 'max_jaccard'), or thr=c('lpt', 'max_sens_spec', 'sensitivity', sens='0.8'), or thr=c('lpt', 'max_sens_spec', 'sensitivity'). Function will use all thresholds if no threshold is specified.
Default "equal_sens_spec".}

\item{buffer}{numeric. Buffer width use in 'bmcp' approach. The buffer width will be interpreted in m if raster used in cont_suit has a longitude/latitude CRS, or map units in other cases. Usage buffer=50000. Default NULL}
}
\value{
# This function return a SpatRaster with continuous and binary prediction.
}
\description{
These methods reduce overprediction of species distribution models based on a posteriori method (see Mendes et al 2020), i.e., the combination of the patterns of species occurrences and predicted suitability
}
\details{
Abbreviation list

\itemize{
\item SDM: species distribution model
\item l: suitability patches that intercept species occurrences
\item k: suitability patches that do not intercept species occurrences
\item T: threshold distances used to select suitability patches
}


These methods reduce overprediction of species distribution models already fitted
based on the occurrences and suitability patterns of species
(see 'thr' arguments)


Method 'obr' (Occurrences based restriction)-
This method assumes that suitable patches intercepting species occurrences (l)
are likely a part of species distributions than suitable patches that do not
intercept any occurrence (k). Distance from all patches k species occurrences to the closest l
patch is calculated, later it is removed k patches that trespass a species-specific
distance threshold from SDMs models. This threshold (T) is calculated as the
maximum distance in a vector of minimal pairwise distances between occurrences.
Whenever a suitable pixel is within a k patch distant from the closest l in less than T,
the suitability of the pixel was reduced to zero. We assumed that this simple threshold
is a surrogate of the species-specific dispersal ability. If T is low, either the species
has been sampled throughout its distribution, or the species is geographically restricted,
justifying a narrow inclusion of k patches (Mendes et al., 2020).

Method 'pres' (Only occurrences based restriction). This is a more restrictive variant of the 'obr' method. It only retains those pixels in suitability patches intercepting occurrences (k) (Mendes et al., 2020).

Method 'lq (Lower Quantile). This method is similar to the 'obr' method, except by the
procedure to define a distance threshold to withdrawn k patches, which is the
lower quartile distance between k patches to the closest l patch. Whenever a suitable
pixel is within a k patch, i.e., not within this lower quartile, the suitability of the
pixel is reduced to zero. This means that 75\% of k patches were withdrawn from the model (Mendes et al., 2020).

Method 'mcp' (Minimum Convex Polygon). Compiled and adapted from
Kremen et al. (2008), this method excludes from SDMs climate suitable
pixels that do not intercept a minimum convex polygon,
with interior angles smaller than 180, enclosing all occurrences of a species.

Method 'bmcp' (Buffered Minimum Convex Polygon). Compiled and adapted
from Kremen et al. (2008), it is similar to the 'mcp' except by the inclusion of a
buffer zone surrounding minimum convex polygons. When used with the "single" options for buffer argument
function will ask for a value in km to be used as the buffer with. When used "species_specific" a buffer will be calculated for each species based on the presences occurrences patterns, assuming as buffer width
the maximum distance in a vector of minimal pairwise distances between occurrences.

Further methodological and performance information of these methods see Mendes et al. (2020).

If used one these constraining method cite Mendes et al 2020.
}
\examples{
\dontrun{
require(dplyr)
require(terra)

data("spp")
somevar <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(somevar)


# It will prepared data for modeling a species
set.seed(10)
occ <- spp \%>\%
  dplyr::filter(species == "sp2") \%>\% # filter a species
  sdm_extract(
    data = ., x = "x", y = "y",
    env_layer = somevar, filter_na = TRUE
  ) \%>\% # extrac variables values
  part_random(.,
    pr_ab = "pr_ab",
    method = c(method = "kfold", folds = 10)
  ) # add columns with partition

# Lets fit and predict a model
m_glm <- fit_glm(
  data = occ,
  response = "pr_ab",
  predictors = names(somevar),
  partition = ".part",
  thr = "equal_sens_spec",
)

# Lets predict this model
m_pred <- sdm_predict(models = m_glm, pred = somevar, thr = NULL, con_thr = FALSE)
plot(m_pred[[1]])
m_pred[[1]] \%>\% plot()


### bmcp method
m_bmcp <- msdm_posteriori(
  records = occ,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  method = "bmcp",
  cont_suit = m_pred[[1]],
  thr = "equal_sens_spec",
  buffer = 30000
)

plot(m_bmcp)


### mcp method
m_mcp <- msdm_posteriori(
  records = occ,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  method = "mcp",
  cont_suit = m_pred[[1]],
  thr = "equal_sens_spec",
  buffer = NULL
)

plot(m_mcp)


### pres method
m_pres <- msdm_posteriori(
  records = occ,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  method = "pres",
  cont_suit = m_pred[[1]],
  thr = "equal_sens_spec",
  buffer = NULL
)

plot(m_pres)


### lq method
m_lq <- msdm_posteriori(
  records = occ,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  method = "lq",
  cont_suit = m_pred[[1]],
  thr = "equal_sens_spec",
  buffer = NULL
)

plot(m_lq)


### obr method
m_obr <- msdm_posteriori(
  records = occ,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  method = "obr",
  cont_suit = m_pred[[1]],
  thr = "equal_sens_spec",
  buffer = NULL
)

plot(m_obr)
}

}
\references{
\itemize{
\item Mendes, P.; Velazco S.J.E.; Andrade, A.F.A.; De Marco, P. (2020) Dealing with overprediction in
species distribution models: how adding distance constraints can improve model accuracy,
Ecological Modelling, in press. https://doi.org/10.1016/j.ecolmodel.2020.109180
\item Kremen, C., Cameron, A., Moilanen, A., Phillips, S. J., Thomas, C. D.,
Beentje, H., . Zjhra, M. L. (2008). Aligning Conservation Priorities Across
Taxa in Madagascar with High-Resolution Planning Tools. Science, 320(5873),
222-226. doi:10.1126/science.1155193
}
}
\seealso{
\code{\link{msdm_priori}}
}
