% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occfilt_select.R
\name{occfilt_select}
\alias{occfilt_select}
\title{Select filtered occurrences}
\usage{
occfilt_select(occ_list, x, y, env_layer, filter_prop = FALSE)
}
\arguments{
\item{occ_list}{list. A list with filtered specie occurrences testing several values (see \code{\link{occfilt_env}} and \code{\link{occfilt_geo}})}

\item{x}{character. Column name with longitude data}

\item{y}{character. Column name with latitude data}

\item{env_layer}{SpatRaster. Raster variables that will be used to fit the model. Factor variables will be removed.}

\item{filter_prop}{logical. If TRUE, the function will return a list with the filtered occurrences and a tibble
with the spatial autocorrelation and number of occurrence values}
}
\value{
If filter_prop = FALSE, a tibble with selected filtered occurrences.
If filter_prop is TRUE, a list with following elemts:
tibble with the spatial autocorrelation and number of occurrence values
\itemize{
  \item A tibble with selected filtered occurrences
  \item A tibble with values used to filter (filt_value), number of occurrence (n_records), and
  mean spatial autocorrelation (mean_autocorr).
  From mean_autocorr column the spatial autocorrelation values for each variable are presented.
  }
}
\description{
Select filtered occurrences based on number of records and spatial autocorrelation (see details)
}
\details{
The function implement the approach used in Velazco et al. (2020)
which consists in calculating for each filtered dataset:
\itemize{
  \item 1- the number of occurrence.
  \item 2- the spatial autocorrelation based on Morans'I for each variable
  \item 3- the mean spatial autocorrelation among variables
  }

Then function will select those dataset with average spatial autocorrelation lower
than the mean of all dataset, and from this subset will select the one with
the highest number occurrences.

If use occfilt_select cite Velazco et al. (2020) as reference.
}
\examples{
\dontrun{
require(terra)
require(dplyr)

# Environmental variables
somevar <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(somevar)

plot(somevar)

# Species occurrences
data("spp")
spp
spp1 <- spp \%>\% dplyr::filter(species == "sp1", pr_ab == 1)

## \%######################################################\%##
####                  Cellsize method                   ####
## \%######################################################\%##
# Using cellsize method
filtered_occ <- occfilt_geo(
  data = spp1,
  x = "x",
  y = "y",
  env_layer = somevar,
  method = c("cellsize", factor = c(1, 4, 8, 12, 16, 20)),
  prj = crs(somevar)
)

filtered_occ

# Select filtered occurrences based on
# number of records and spatial autocorrelation
occ_selected <- occfilt_select(
  occ_list = filtered_occ,
  x = "x",
  y = "y",
  env_layer = somevar,
  filter_prop = FALSE
)
occ_selected

occ_selected <- occfilt_select(
  occ_list = filtered_occ,
  x = "x",
  y = "y",
  env_layer = somevar,
  filter_prop = TRUE
)
occ_selected$occ

occ_selected$filter_prop
}
}
\references{
\itemize{
\item Velazco, S. J. E., Svenning, J-C., Ribeiro, B. R., & Laureto, L. M. O. (2020). On
opportunities and threats to conserve the phylogenetic diversity of Neotropical palms.
Diversity and Distributions, 27, 512–523. https://doi.org/10.1111/ddi.13215
}
}
\seealso{
\code{\link{occfilt_env}}, \code{\link{occfilt_geo}}
}
