% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_sband.R
\name{part_sband}
\alias{part_sband}
\title{Spatial band cross-validation}
\usage{
part_sband(
  env_layer,
  data,
  x,
  y,
  pr_ab,
  type = "lon",
  n_part = 2,
  min_bands = 2,
  max_bands = 20,
  prop = 0.5
)
}
\arguments{
\item{env_layer}{SpatRaster. Raster with environmental
variable. This will be used to evaluate spatial autocorrelation and
environmental similarity between training and testing partition. Because this function
calculate dissimilarity based on euclidean distances, it can only handle continuous
layers, do not use categorical layers as inputs}

\item{data}{data.frame. Data.frame or tibble object with presences
(or presence-absence, or presences-pseudo-absence) records, and coordinates}

\item{x}{character. Column name with longitude data}

\item{y}{character. Column name with latitude data}

\item{pr_ab}{character. Column with presences, presence-absence,
or pseudo-absence. Presences must be represented by 1 and absences by 0}

\item{type}{character. Specify bands across different degrees of longitude 'lon' or latitude 'lat'. Default is 'lon'.}

\item{n_part}{integer. Number of partition. Default 2, values other than
2 has not yet been implemented.}

\item{min_bands}{integer. Minimum number of spatial bands to be tested, default 2.}

\item{max_bands}{integer. Maximum number of spatial bands to be tested, default 20.}

\item{prop}{numeric. Proportion of point used for testing autocorrelation between
groups (values > 0 and <=1). The smaller this number is, the faster the function will work.
Default 0.5}
}
\value{
A list with:
\itemize{
  \item part: A tibble object with information used in 'data' arguments and a additional column .part with partition group.
  \item best_part_info: A tibble with information of the bets partition. It contains the number of the best partition (n_grid), number of bands (n_bands), standard deviation of presences (sd_p), standard deviation of absences (sd_a), Moran's I spatial autocorrelation (spa_auto), and environmental similarity based on euclidean distance (env_sim).
  \item grid: A SpatRaster object with bands
  }
}
\description{
This function explores different numbers of spatial bands and returns the
best one suited for a given presence or presence-absences database. The selection of the best
number of bands is performed automatically considering spatial autocorrelation, environmental
similarity, and the number of presence and absence records in each partition.
}
