% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_senv.R
\name{part_senv}
\alias{part_senv}
\title{Environmental and spatial cross-validation}
\usage{
part_senv(
  env_layer,
  data,
  x,
  y,
  pr_ab,
  min_n_groups = 2,
  max_n_groups = 10,
  min_occ = 10,
  prop = 0.5
)
}
\arguments{
\item{env_layer}{SpatRaster. Raster with environmental
variable. This will be used to evaluate spatial autocorrelation and
environmental similarity between training and testing partitions. Because this function
calculate dissimilarity based on Euclidean distances, it can only be used with continuous variables}

\item{data}{data.frame. Data.frame or tibble object with presence
(or presence-absence, or presences-pseudo-absence) records, and coordinates}

\item{x}{character. Column name with spatial x coordinates}

\item{y}{character. Column name with spatial y coordinates}

\item{pr_ab}{character. Column with presences, presence-absence,
or pseudo-absence. Presences must be represented by 1 and absences by 0}

\item{min_n_groups}{integer. Minimum number of groups to be tested. Default 2.}

\item{max_n_groups}{integer. Maximum number of groups to be tested. Default 10.}

\item{min_occ}{numeric. Minimum number of presences or absences in a partition fold.
The min_occ value should be base on the amount of predictors in order to avoid over-fitting
or error when fitting models for a given fold. Default 10.}

\item{prop}{numeric. Proportion of point used for testing autocorrelation between
groups (values > 0 and <=1). The smaller this number is, the faster the function will work.
Default 0.5}
}
\value{
A list with:
\itemize{
  \item part: A tibble object with information used in 'data' arguments and a additional column .part with partition group.
  \item best_part_info: A tibble with information about the best partition. It contains the number of partition (n_groups), standard deviation of presences (sd_p), standard deviation of absences (sd_a), Moran's I spatial autocorrelation (spa_auto) and environmental similarity based on Euclidean distance (env_sim)
  }
}
\description{
This function explores different numbers of environmental partitions based on the
K-means clustering algorithm and returns the number of partitions best suited for a given presence or
presence-absences database. Selection of the best number of partitions is performed
automatically considering spatial autocorrelation, environmental similarity, and the
number of presence and/or absence records in each partition.
}
\details{
write here criteria used for performing the search of the best partition (metrics and quartile selection).
}
\examples{
\dontrun{
}
}
\seealso{
\code{\link{part_random}}, and \code{\link{part_sblock}}.
}
