% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_senv.R
\name{part_senv}
\alias{part_senv}
\title{Environmental and spatial cross validation}
\usage{
part_senv(
  env_layer,
  data,
  x,
  y,
  pr_ab,
  min_n_groups = 2,
  max_n_groups = 10,
  prop = 0.5
)
}
\arguments{
\item{env_layer}{SpatRaster. Raster with environmental
variable. This will be used to evaluate spatial autocorrelation and
environmental similarity between training and testing partition. Because this function
calculate dissimilarity based on euclidean distances, it can only handle continuous
layers, do not use categorical layers as inputs}

\item{data}{data.frame. Data.frame or tibble object with presences
(or presence-absence, or presences-pseudo-absence) records, and coordinates}

\item{x}{character. Column name with longitude data}

\item{y}{character. Column name with latitude data}

\item{pr_ab}{character. Column with presences, presence-absence,
or pseudo-absence. Presences must be represented by 1 and absences by 0}

\item{min_n_groups}{integer. Minimum number of groups to be tested. Default 2.}

\item{max_n_groups}{integer. Maximum number of groups to be tested. Default 10.}

\item{prop}{numeric. Proportion of point used for testing autocorrelation between
groups (values > 0 and <=1). The smaller this number is, the faster the function will work.
Default 0.5}
}
\value{
A list with:
\itemize{
  \item part: A tibble object with information used in 'data' arguments and a additional column .part with partition group.
  \item best_part_info: A tibble with information of the bets partition. It contains the number of partition (n_groups), standard deviation of presences (sd_p), standard deviation of absences (sd_a), Moran's I spatial autocorrelation (spa_auto) and environmental similarity based on euclidean distance (env_sim)
  }
}
\description{
This function explores different numbers of environmental partitions based on the
K-mean cluster algorithm and returns the best one suited for a given presence or
presence-absences database. The selection of the best number of partition is performed
automatically considering spatial autocorrelation, environmental similarity, and the
number of presence and/or absence records in each partition.
}
\details{
write here criteria used for performing the search of the best partition (metrics and quartil selection).
}
\examples{
\dontrun{
require(terra)
require(ggplot2)

f <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(f)

# Select a species
spp1 <- spp \%>\% dplyr::filter(species == "sp1")

part1 <- part_senv(
  env_layer = somevar,
  data = spp1,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  min_n_groups = 2,
  max_n_groups = 10,
  prop = 0.2
)

part1

ggplot(part1$part, aes(x, y, col = factor(.part))) +
  geom_point(aes(shape = factor(pr_ab)))

ggplot(part1$part, aes(x, y, col = factor(.part))) +
  geom_point(aes(shape = factor(pr_ab))) +
  facet_wrap(. ~ .part)

ggplot(part1$part, aes(x, y, col = factor(.part))) +
  geom_point(aes(shape = factor(pr_ab))) +
  facet_wrap(. ~ pr_ab)
}
}
\seealso{
\code{\link{part_classical}}, and \code{\link{part_sblock}}.
}
