% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res.R
\name{plot_res}
\alias{plot_res}
\title{Plot the resolutions to be used in block_partition_pa}
\usage{
plot_res(r, max_res_mult)
}
\arguments{
\item{r}{raster. A raster layer. Preferably a layer of environmental variables to be used}

\item{res_mult}{numeric. Maximum resolution to be tested.}
}
\value{
A plot with raster and grid with the resolution setted
}
\description{
Plot the resolutions to be used in block_partition_pa
}
\examples{
\dontrun{
f <- system.file("external/test.grd", package = "raster")
f
r <- raster(f)
plot_res(r, max_res_mult = 2)
plot_res(r, max_res_mult = 10)
plot_res(r, max_res_mult = 20)
}
}
