% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_background.R
\name{sample_background}
\alias{sample_background}
\title{Sample background points}
\usage{
sample_background(n, rlayer, maskval = NULL)
}
\arguments{
\item{n}{integer. Number of background point to be sampled}

\item{rlayer}{raster. A raster layer used for sampling background-point.
It is recommended to use a layer with the same resolution and extent that environmental variables that will be used for modeling. In the case use maskval argument, this raster layer must contain the values to sampling constraint}

\item{maskval}{integer or numeric. Values of the raster layer used for constraining the background points sampling}
}
\value{

}
\description{
Sample background points
}
\examples{
\dontrun{
require(dplyr)
data(spp)
data(somevar)

# Lest practice with a single species
single_spp <- spp \%>\% dplyr::filter(species == "sp3")
part <- block_partition(
  env_layer = somevar,
  data = single_spp,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  min_res_mult = 100,
  max_res_mult = 500,
  num_grids = 10,
  n_part = 2
)

grid_env <- get_block(env_layer = somevar, best_grid = part$Grid)
plot(grid_env)


# Sample background points throughout study area
bg <- sample_background(n = 1000, rlayer = grid_env)
bg \%>\% plot()

# Sample background points constrained to a region regions with a give set of values
sample_background(n = 1000, rlayer = grid_env, maskval = 1) \%>\% plot()
sample_background(n = 1000, rlayer = grid_env, maskval = 2) \%>\% plot()
sample_background(n = 1000, rlayer = grid_env, maskval = c(1, 2)) \%>\% plot()
}

}
