% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_background.R
\name{sample_background}
\alias{sample_background}
\title{Sample background points}
\usage{
sample_background(n, rlayer, maskval = NULL, calibarea = NULL)
}
\arguments{
\item{n}{integer. Number of background point to be sampled}

\item{rlayer}{raster. A raster layer used for sampling background-point.
It is recommended to use a layer with the same resolution and extent that environmental variables that will be used for modeling. In the case use maskval argument, this raster layer must contain the values to sampling constraint}

\item{maskval}{integer or numeric. Values of the raster layer used for constraining the background points sampling}

\item{calibarea}{shapefile. A SpatialPolygon or SpatialPolygonDataFrame which delimit the calibration area used for a given species (see calib_area function).}
}
\value{

}
\description{
Sample background points
}
\examples{
\dontrun{
require(dplyr)
data(spp)
data(somevar)

# Lest practice with a single species
single_spp <- spp \%>\% dplyr::filter(species == "sp3")

part <- block_partition(
  env_layer = somevar,
  data = single_spp,
  x = "x",
  y = "y",
  pr_ab = "pr_ab",
  min_res_mult = 100,
  max_res_mult = 500,
  num_grids = 10,
  n_part = 2
)

grid_env <- get_block(env_layer = somevar, best_grid = part$Grid)
plot(grid_env)

# Sample background points throughout study area
bg <- sample_background(n = 1000, rlayer = grid_env)
plot(grid_env)
points(bg)

# Sample background points constrained to a region regions with a give set of values
plot(grid_env)
sample_background(n = 1000, rlayer = grid_env, maskval = 1) \%>\% points()
plot(grid_env)
sample_background(n = 1000, rlayer = grid_env, maskval = 2) \%>\% points()
plot(grid_env)
sample_background(n = 1000, rlayer = grid_env, maskval = c(1, 2)) \%>\% points()

# Sample background within a calibration area and constrained to a region regions
# Delimit a calibration area with calib_area
ca_ps1 <- calib_area(
  data = single_spp,
  x = 'x',
  y = 'y',
  method = c('buffer', width=50000),
)
plot(grid_env)
plot(ca_ps1, add=T)
points(single_spp[-1], col='blue', cex=0.7, pch=19)
sample_background(n = 1000, rlayer = grid_env, maskval = 1, calibarea = ca_ps1) \%>\%
  points(col='red')

}

}
\seealso{
\code{\link{sample_pseudoabs}} and \code{\link{calib_area}}.
}
