% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_directory.R
\name{sdm_directory}
\alias{sdm_directory}
\title{Create directories for saving the outputs of the flexsdm workflow}
\usage{
sdm_directory(
  main_dir,
  proj_dir = FALSE,
  save_results = FALSE,
  models = NULL,
  ensemble = NULL,
  msdm = NULL
)
}
\arguments{
\item{main_dir}{character. Directory path containing main folder for saving model inputs and outputs}

\item{proj_dir}{logical. If TRUE, function will create a folder for predictors for different regions or time periods used to project models. (default FALSE).}

\item{save_results}{logical. If TRUE, function will create a folder for model results. (default FALSE).}

\item{models}{vector. Vector of model types that will be used. (default NULL)}

\item{ensemble}{vector. Vector of methods used to ensemble different models. (default NULL)}

\item{msdm}{logical. If TRUE, function will create a folder for the output of two methods for correcting SDM overprediction (see msdm_priori and msdm_posteriori). (default FALSE)}
}
\value{
None
}
\description{
Create directories for saving the outputs of the flexsdm workflow
}
\examples{

# Create directory for SDM workflow
my_dir <- file.path(getwd(), 'flexsdm_example')
dir.create(my_dir)

# Implement sdm_directory
sdm_directory(main_dir = my_dir,
proj_dir = TRUE,
save_results = TRUE,
models = c('GLM', 'GAM', 'RANDOM_FOREST'),
ensemble = c('MEAN', "MEANSUP"),
msdm = TRUE)




}
