% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_eval.R
\name{sdm_eval}
\alias{sdm_eval}
\title{Calculate different model performance metrics}
\usage{
sdm_eval(p, a, bg = NULL, thr = NULL)
}
\arguments{
\item{p}{numeric. Predicted suitability for presences}

\item{a}{numeric. Predicted suitability for presences absences}

\item{bg}{numeric. Predicted suitability for background points. It is used for BOYCE metric.
It bg is set as null BOYCE metric will be calculated with presences and absences suitability
values}

\item{thr}{character. Threshold used to get binary suitability values (i.e. 0,1).
It is useful for threshold-dependent performance metrics.
It is possible to use more than one threshold type.
It is necessary to provide a vector for this argument. The next threshold area available:
\itemize{
  \item lpt: The highest threshold at which there is no omission.
  \item equal_sens_spec: Threshold at which the sensitivity and specificity are equal.
  \item max_sens_spec: Threshold at which the sum of the sensitivity and specificity
  is the highest (aka threshold that maximizes the TSS).
  \item max_jaccard: The threshold at which Jaccard is the highest.
  \item max_sorensen: The threshold at which Sorensen is highest.
  \item max_fpb: The threshold at which FPB is highest.
  \item sensitivity: Threshold based on a specified sensitivity value.
  Usage thr = c('sensitivity', sens='0.6') or thr = c('sensitivity'). 'sens' refers
  to sensitivity value. If it is not specified a sensitivity values, function will use
  by default 0.9
  }
In the case of use more than one threshold type it is necessary concatenate threshold types,
e.g., thr=c('lpt', 'max_sens_spec', 'max_jaccard'), or thr=c('lpt', 'max_sens_spec',
'sensitivity', sens='0.8'), or thr=c('lpt', 'max_sens_spec', 'sensitivity').
Function will use all thresholds if no threshold is specified}
}
\value{
a tibble with next columns
\itemize{
\item threshold: threshold names.
\item thr_value: threshold value
\item n_presences: threshold names.
\item n_absences: threshold names.
\item from TPR to IMAE: performance metrics
}
}
\description{
This function calculates threshold dependent and independent performance metric.
It calculates TPR (True Positive Rate, also called sensitivity), TNR (True Negative Rate,
also called specificity), SORENSEN, JACCARD, FPB, OR (Omission Rate), TSS (True Skill Statistic)
KAPPA, AUC (Area Under Curve), BOYCE, IMAE (Inverse Mean Absolute Error, i.e.
1-Mean Absolute Error).
}
\examples{
\dontrun{
require(dplyr)

set.seed(0)
p <- rnorm(50, mean = 0.7, sd = 0.3) \%>\% abs()
p[p > 1] <- 1
p[p < 0] <- 0

set.seed(0)
a <- rnorm(50, mean = 0.3, sd = 0.2) \%>\% abs()
a[a > 1] <- 1
a[a < 0] <- 0

set.seed(0)
backg <- rnorm(1000, mean = 0.4, sd = 0.4) \%>\% abs()
backg[backg > 1] <- 1
backg[backg < 0] <- 0

# Function use without threshold specification
e <- sdm_eval(p, a)
e

# Function use with threshold specification
sdm_eval(p, a, thr = "max_sorensen")
sdm_eval(p, a, thr = c("lpt", "max_sens_spec", "max_jaccard"))
sdm_eval(p, a, thr = c("lpt", "max_sens_spec", "sensitivity"))
sdm_eval(p, a, thr = c("lpt", "max_sens_spec", "sensitivity", sens = "0.95"))

# Use of bg argument (it will only be used for calculating BOYCE index)
sdm_eval(p, a, thr = "max_sens_spec")
sdm_eval(p, a, thr = c("max_sens_spec"), bg = backg)

# I the case it is needed use background for calculate all other metric
# background values can be used in "a" argument
sdm_eval(p, backg, thr = "max_sens_spec")
}

}
