% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_extract.R
\name{sdm_extract}
\alias{sdm_extract}
\title{Extract environmental data based on x and y coordinates}
\usage{
sdm_extract(data, x, y, predictors, env_layer, filter_na = TRUE)
}
\arguments{
\item{data}{data.frame. Database with presences, presence-absence, or pseudo-absence, records with x and y coordinates}

\item{x}{character. Column name with longitude data}

\item{y}{character. Column name with latitude data}

\item{predictors}{character. Vector with the variable names of predictor variables
Usage predictors = c("aet", "cwd", "tmin")}

\item{env_layer}{SpatRaster. Raster with environmental variables.}

\item{filter_na}{logical. If filter_na = TRUE (default), the rows with NA values for any of the
environmental variables are removed from the returned tibble.}
}
\value{
A tibble that returns the original data base of presence, presence-absence, or pseudo-absence location records with additional columns
for the extracted environmental variables at each xy location from the SpatRast object 'env_layer'
}
\description{
Extract environmental data based on x and y coordinates
}
\examples{
\dontrun{
require(terra)

# Load datasets
data(spp)
f <- system.file("external/somevar.tif", package = "flexsdm")
somevar <- terra::rast(f)

# Extract environmental data from somevar for locations in spp
ex_spp <-
  sdm_extract(
    data = spp,
    x = "x",
    y = "y",
    predictors = names(somevar),
    env_layer = some_var,
    filter_na = TRUE
  )

view(ex_spp)
}
}
