% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_predict.R
\name{sdm_predict}
\alias{sdm_predict}
\title{Spatial model prediction}
\usage{
sdm_predict(
  models,
  pred,
  thr = NULL,
  con_thr = FALSE,
  calib_area = NULL,
  clamp = TRUE,
  pred_type = "cloglog"
)
}
\arguments{
\item{models}{list. A list a single or several models fitted with some of fit_ or tune_ functions}

\item{pred}{SpatRaster. Raster layer with predictor variables. Names of layers must exactly match those used in model fitting.}

\item{thr}{character. Threshold used to get binary suitability values (i.e. 0,1). It is possible to use more than one threshold type. It is mandatory to use the same threshold/s used to fit the models. The next threshold area available:
\itemize{
  \item lpt: The highest threshold at which there is no omission.
  \item equal_sens_spec: Threshold at which the sensitivity and specificity are equal.
  \item max_sens_spec: Threshold at which the sum of the sensitivity and specificity is the highest (aka threshold that maximizes the TSS).
  \item max_jaccard: The threshold at which Jaccard is the highest.
  \item max_sorensen: The threshold at which Sorensen is highest.
  \item max_fpb: The threshold at which FPB is highest.
  \item sensitivity: Threshold based on a specified sensitivity value used to fit the models.
  \item all: Will be used all the threshold filled in the 'models' argument. It is not possible to combine with previews one
  }
Usage thr = c('lpt', 'max_sens_spec', 'max_jaccard'), thr=c('lpt', 'max_sens_spec', 'sensitivity'), or thr='all'. If no threshold is specified (i.e., thr = NULL) function will return continuous prediction only. Default NULL}

\item{con_thr}{logical. If true will be returned predictions with suitability values above threshold/s. Default = FALSE}

\item{calib_area}{SpatialPolygon or SpatialPolygonDataFrame. Spatial polygon used for restring prediction into a given region. Default = NULL}

\item{clamp}{logical. It is set with TRUE, predictors and features are restricted to the range seen during model training. Only valid for Maxent model (see tune_mx and fit_mx)}

\item{pred_type}{character. Type of response required available "link", "exponential", "cloglog" and "logistic". Default "cloglog". Only valid for Maxent model (see tune_mx and fit_mx)}
}
\value{
A list of SpatRaster with continuous or continuous and binary predictions
}
\description{
This function allows the prediction of one or more models built with the fit_ or tune_ function set. It can return continuous or continuous and binary predictions for one or more thresholds
}
