% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_summarize.R
\name{sdm_summarize}
\alias{sdm_summarize}
\title{Merge model performance tables}
\usage{
sdm_summarize(models)
}
\arguments{
\item{models}{}
}
\value{

}
\description{
Merge model performance tables
}
\examples{
\dontrun{
models <- list(m_1, m_2, m_3)
}
}
