% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_gbm.R
\name{tune_gbm}
\alias{tune_gbm}
\title{Function for constructing Random Forest with exploration of hyper-parameters}
\usage{
tune_gbm(
  data,
  response,
  predictors,
  predictors_f = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  ...
)
}
\arguments{
\item{data}{}

\item{response}{}

\item{predictors}{}

\item{predictors_f}{}

\item{partition}{}

\item{grid}{}

\item{thr}{}

\item{metric}{}
}
\value{

}
\description{
Function for constructing Random Forest with exploration of hyper-parameters
}
