% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_mx.R
\name{tune_mx}
\alias{tune_mx}
\title{Fit and validate Maximum Entropy models with exploration of hyper-parameters}
\usage{
tune_mx(
  data,
  response,
  predictors,
  predictors_f = NULL,
  background = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  clamp = TRUE,
  pred_type = "cloglog",
  ...
)
}
\arguments{
\item{clamp}{logical. It is set with TRUE, predictors and features are restricted to the range seen during model training.}

\item{pred_type}{character. Type of response required available "link", "exponential", "cloglog" and "logistic". Default "cloglog"}
}
\value{

}
\description{
Fit and validate Maximum Entropy models with exploration of hyper-parameters
}
