% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_svm.R
\name{tune_svm}
\alias{tune_svm}
\title{Function for constructing Support Vector Machine with exploration of hyper-parameters}
\usage{
tune_svm(
  data,
  response,
  predictors,
  predictors_f = NULL,
  partition,
  grid = NULL,
  thr = NULL,
  metric = "TSS",
  ...
)
}
\arguments{
\item{metric}{}
}
\value{

}
\description{
Function for constructing Support Vector Machine with exploration of hyper-parameters
}
